requirements.txt
dash==2.7.0
dash_bootstrap_components==1.4.2
matplotlib==3.6.2
scipy==1.9.3
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
scikit_learn==1.1.3
umap_learn==0.5.3
gunicorn==21.2.0
